requirements.txt
flask
requests
Pillow

